+++ akshare/ws/js_ws_news.py
@@ -9,7 +9,7 @@ import pandas as pd
 import requests
 
 
-def js_news(timestamp: str = "2021-06-05 20:50:18") -> pd.DataFrame:
+def js_news(timestamp: str = "2022-11-13 14:50:18") -> pd.DataFrame:
     """
     金十数据-市场快讯
     https://www.jin10.com/
@@ -22,7 +22,7 @@ def js_news(timestamp: str = "2021-06-05 20:50:18") -> pd.DataFrame:
     params = {
         "channel": "-8200",
         "vip": "1",
-        "t": "1625623640730",
+        "t": "1668329215869",
         "max_time": timestamp,
     }
     headers = {
@@ -30,12 +30,12 @@ def js_news(timestamp: str = "2021-06-05 20:50:18") -> pd.DataFrame:
         "accept-encoding": "gzip, deflate, br",
         "accept-language": "zh-CN,zh;q=0.9,en;q=0.8",
         "cache-control": "no-cache",
-        "handleerror": "true",
         "origin": "https://www.jin10.com",
         "pragma": "no-cache",
         "referer": "https://www.jin10.com/",
         "sec-ch-ua": '" Not;A Brand";v="99", "Google Chrome";v="91", "Chromium";v="91"',
         "sec-ch-ua-mobile": "?0",
+        'handleerror': 'true',
         "sec-fetch-dest": "empty",
         "sec-fetch-mode": "cors",
         "sec-fetch-site": "same-site",
@@ -62,5 +62,5 @@ def js_news(timestamp: str = "2021-06-05 20:50:18") -> pd.DataFrame:
 
 
 if __name__ == "__main__":
-    js_news_df = js_news(timestamp="2022-07-13 17:20:18")
+    js_news_df = js_news(timestamp="2022-11-13 12:40:25")
     print(js_news_df)