+++ docs/changelog.md
@@ -35,6 +35,14 @@
 
 ## 更新说明
 
+1.7.92 fix: fix fund_portfolio_hold_em interface
+
+    1. 修复 fund_portfolio_hold_em 接口
+
+1.7.91 fix: fix futures_price_index_nh interface
+
+    1. 修复 futures_price_index_nh 接口时区对齐问题
+
 1.7.90 fix: fix stock_yjbb_em interface
 
     1. 修复 stock_yjbb_em 接口
@@ -1476,6 +1484,10 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.7.92 fix: fix fund_portfolio_hold_em interface
+
+1.7.91 fix: fix futures_price_index_nh interface
 
 1.7.90 fix: fix stock_yjbb_em interface